+++ NEWS.md
@@ -1,6 +1,6 @@
 ## usefunc 1.1.6 2024_11_19
 
-* Add `blend_colours()` function
+* Add `blend_palette()` function
 * Add {monochromeR} to Imports
 
 ## usefunc 1.1.5 2024_11_17